refactor: 순수 도메인 레포지토리를 domain 패키지로 이동
- picks: PicksRepository 이동 및 import 경로 수정
- support/block: UserBlockRepository 이동 및 import 경로 수정
- support/help: HelpRepository 이동 및 import 경로 수정
- support/business: BusinessSupportRepository 이동 및 import 경로 수정
- JPA 레포지토리에 @JpaRepositoryImpl 어노테이션 적용
- CLAUDE.md 가이드라인 준수: 순수 레포지토리는 domain, JPA 구현체는 repository 패키지
