+++ bottlenote-mono/src/main/java/app/bottlenote/picks/repository/JpaPicksRepository.java
@@ -2,6 +2,7 @@ package app.bottlenote.picks.repository;
 
 import app.bottlenote.common.annotation.JpaRepositoryImpl;
 import app.bottlenote.picks.domain.Picks;
+import app.bottlenote.picks.domain.PicksRepository;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
