+++ bottlenote-mono/src/main/java/app/bottlenote/picks/service/PicksCommandService.java
@@ -9,10 +9,10 @@ import app.bottlenote.alcohols.facade.AlcoholFacade;
 import app.bottlenote.history.event.publisher.HistoryEventPublisher;
 import app.bottlenote.picks.constant.PicksStatus;
 import app.bottlenote.picks.domain.Picks;
+import app.bottlenote.picks.domain.PicksRepository;
 import app.bottlenote.picks.dto.request.PicksUpdateRequest;
 import app.bottlenote.picks.dto.response.PicksUpdateResponse;
 import app.bottlenote.picks.event.payload.PicksRegistryEvent;
-import app.bottlenote.picks.repository.PicksRepository;
 import app.bottlenote.user.exception.UserException;
 import app.bottlenote.user.facade.UserFacade;
 import lombok.RequiredArgsConstructor;
