+++ bottlenote-mono/src/main/java/app/bottlenote/support/block/domain/UserBlockRepository.java
@@ -1,4 +1,4 @@
-package app.bottlenote.support.block.repository;
+package app.bottlenote.support.block.domain;
 
 import app.bottlenote.support.block.domain.UserBlock;
 import app.bottlenote.support.block.dto.response.UserBlockItem;
