+++ bottlenote-mono/src/main/java/app/bottlenote/support/block/repository/JpaUserBlockRepository.java
@@ -1,6 +1,8 @@
 package app.bottlenote.support.block.repository;
 
+import app.bottlenote.common.annotation.JpaRepositoryImpl;
 import app.bottlenote.support.block.domain.UserBlock;
+import app.bottlenote.support.block.domain.UserBlockRepository;
 import app.bottlenote.support.block.dto.response.UserBlockItem;
 import java.util.List;
 import java.util.Optional;
@@ -8,9 +10,8 @@ import java.util.Set;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
-import org.springframework.stereotype.Repository;
 
-@Repository
+@JpaRepositoryImpl
 public interface JpaUserBlockRepository
     extends UserBlockRepository, JpaRepository<UserBlock, Long> {
