+++ bottlenote-mono/src/main/java/app/bottlenote/support/business/domain/BusinessSupportRepository.java
@@ -1,4 +1,4 @@
-package app.bottlenote.support.business.repository;
+package app.bottlenote.support.business.domain;
 
 import app.bottlenote.support.business.domain.BusinessSupport;
 import java.util.List;
