+++ bottlenote-mono/src/main/java/app/bottlenote/support/help/service/HelpService.java
@@ -8,6 +8,7 @@ import static app.bottlenote.support.help.exception.HelpExceptionCode.HELP_NOT_F
 
 import app.bottlenote.global.service.cursor.PageResponse;
 import app.bottlenote.support.help.domain.Help;
+import app.bottlenote.support.help.domain.HelpRepository;
 import app.bottlenote.support.help.dto.request.HelpImageItem;
 import app.bottlenote.support.help.dto.request.HelpPageableRequest;
 import app.bottlenote.support.help.dto.request.HelpUpsertRequest;
@@ -15,7 +16,6 @@ import app.bottlenote.support.help.dto.response.HelpDetailItem;
 import app.bottlenote.support.help.dto.response.HelpListResponse;
 import app.bottlenote.support.help.dto.response.HelpResultResponse;
 import app.bottlenote.support.help.exception.HelpException;
-import app.bottlenote.support.help.repository.HelpRepository;
 import app.bottlenote.user.facade.UserFacade;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
